feat(utils): update button_error class to include source error message
The button_error class has been updated to include the source error message
when it is provided. This was achieved by modifying the constructor to append
the source error message to the original error message, separated by a comma.
This change improves the debug information available in case of errors.