feat: add bump checking in message create event
Add a new `checkBump` function in the `MessageCreate` class to check for bumps in guilds. If the message is from the bump bot and is in the bump channel, it fetches the interaction and checks if it is the bump command. If it is, it creates a new bump in the database and sends a reply with the bump count and XP earned. It also adds XP to the user if the bump count is less than the max bumps set in the global config.