feat(command.manager.ts): add support for setting command IDs
The changes in this commit add support for setting command IDs in the `CommandManager` class. This is done by introducing a new private method called `setIds`, which takes an array of `ApplicationCommand` objects and updates the corresponding `CommandRun` objects with their respective IDs. This allows for easier tracking and management of commands within the `CommandManager` class.